+++ packages/demo/src/components/examples/BadgeExamples.tsx
@@ -12,6 +12,12 @@ export class BadgeExamples extends React.Component<any, any> {
                     </div>
                 </div>
                 <div className="form-group">
+                    <label className="form-control-label">Badge with Icon</label>
+                    <div className="form-control">
+                        <Badge label="Badge label" icon="lock" />
+                    </div>
+                </div>
+                <div className="form-group">
                     <label className="form-control-label">Badge with extra classes</label>
                     <div className="form-control">
                         <Badge label="Badge blue background" extraClasses={['bg-blue']} />